Moved all generation related logic into `generator.go`; Abstracted behind `Generator` interface for easy mocking in tests
